+++ packages/css-color-parser/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changes to CSS Color Parser
 
+### Unreleased (patch)
+
+- Fix missing rectangular color spaces for interpolation in `color-mix()`.
+- Fix `none` keyword conversion in relative color syntax origin values.
+
 ### 1.6.2
 
 _March 16, 2024_